Move num_shortest_path_unweighted to its own file
