+++ src/shortest_path/mod.rs
@@ -16,6 +16,7 @@ mod astar;
 mod dijkstra;
 mod floyd_warshall;
 mod k_shortest_path;
+mod num_shortest_path;
 
 use hashbrown::{HashMap, HashSet};
 
@@ -28,11 +29,10 @@ use pyo3::Python;
 
 use petgraph::graph::NodeIndex;
 use petgraph::prelude::*;
-use petgraph::visit::{Bfs, EdgeIndexable, NodeCount, NodeIndexable};
+use petgraph::visit::{EdgeIndexable, NodeCount};
 use petgraph::EdgeType;
 
 use ndarray::prelude::*;
-use num_bigint::{BigUint, ToBigUint};
 use numpy::IntoPyArray;
 use rayon::prelude::*;
 
@@ -1190,58 +1190,6 @@ pub fn digraph_floyd_warshall_numpy(
     Ok(mat.into_pyarray(py).into())
 }
 
-fn _num_shortest_paths_unweighted<Ty: EdgeType>(
-    graph: &StableGraph<PyObject, PyObject, Ty>,
-    source: usize,
-) -> PyResult<HashMap<usize, BigUint>> {
-    let mut out_map: Vec<BigUint> =
-        vec![0.to_biguint().unwrap(); graph.node_bound()];
-    let node_index = NodeIndex::new(source);
-    if graph.node_weight(node_index).is_none() {
-        return Err(PyIndexError::new_err(format!(
-            "No node found for index {}",
-            source
-        )));
-    }
-    let mut bfs = Bfs::new(&graph, node_index);
-    let mut distance: Vec<Option<usize>> = vec![None; graph.node_bound()];
-    distance[node_index.index()] = Some(0);
-    out_map[source] = 1.to_biguint().unwrap();
-    while let Some(current) = bfs.next(graph) {
-        let dist_plus_one = distance[current.index()].unwrap_or_default() + 1;
-        let count_current = out_map[current.index()].clone();
-        for neighbor_index in
-            graph.neighbors_directed(current, petgraph::Direction::Outgoing)
-        {
-            let neighbor: usize = neighbor_index.index();
-            if distance[neighbor].is_none() {
-                distance[neighbor] = Some(dist_plus_one);
-                out_map[neighbor] = count_current.clone();
-            } else if distance[neighbor] == Some(dist_plus_one) {
-                out_map[neighbor] += &count_current;
-            }
-        }
-    }
-
-    // Do not count paths to source in output
-    distance[source] = None;
-    out_map[source] = 0.to_biguint().unwrap();
-
-    // Return only nodes that are reachable in the graph
-    Ok(out_map
-        .into_iter()
-        .zip(distance.iter())
-        .enumerate()
-        .filter_map(|(index, (count, dist))| {
-            if dist.is_some() {
-                Some((index, count))
-            } else {
-                None
-            }
-        })
-        .collect())
-}
-
 /// Get the number of unweighted shortest paths from a source node
 ///
 /// :param PyDiGraph graph: The graph to find the number of shortest paths on
@@ -1258,7 +1206,10 @@ pub fn digraph_num_shortest_paths_unweighted(
     source: usize,
 ) -> PyResult<NodesCountMapping> {
     Ok(NodesCountMapping {
-        map: _num_shortest_paths_unweighted(&graph.graph, source)?,
+        map: num_shortest_path::num_shortest_paths_unweighted(
+            &graph.graph,
+            source,
+        )?,
     })
 }
 
@@ -1278,7 +1229,10 @@ pub fn graph_num_shortest_paths_unweighted(
     source: usize,
 ) -> PyResult<NodesCountMapping> {
     Ok(NodesCountMapping {
-        map: _num_shortest_paths_unweighted(&graph.graph, source)?,
+        map: num_shortest_path::num_shortest_paths_unweighted(
+            &graph.graph,
+            source,
+        )?,
     })
 }
